+++ packages/react-svg-core/src/index.ts
@@ -9,20 +9,28 @@ export function optimize(opts: any = {}): (content: string) => Promise<string> {
   opts = validateAndFix(opts);
   const svgo = new Svgo(opts);
 
-  return (content: string) => svgo.optimize(content).then(data => data.data);
+  return (content: string) => svgo.optimize(content).then((data) => data.data);
 }
 
+type TransformOpts = { jsx?: boolean; componentName?: string };
+
 // Babel Transform
-export function transform({ jsx = false }: { jsx?: boolean } = {}): (
-  content: string
-) => string {
-  return content =>
+export function transform(
+  opts: TransformOpts = {}
+): (content: string) => string {
+  const jsx = opts.jsx || false;
+  const componentName = opts.componentName || null;
+
+  return (content) =>
     babelTransform(content, {
       babelrc: false,
       configFile: false,
       presets: [jsx ? void 0 : require.resolve("@babel/preset-react")].filter(
         Boolean
       ),
-      plugins: [require.resolve("@babel/plugin-syntax-jsx"), plugin]
+      plugins: [
+        require.resolve("@babel/plugin-syntax-jsx"),
+        [plugin, { componentName }],
+      ],
     });
 }