+++ packages/rollup-plugin-react-svg/src/index.ts
@@ -8,6 +8,7 @@ type PluginOpts = {
   exclude?: any;
   svgo?: any;
   jsx?: boolean;
+  componentName?: (filePath: string) => string;
 };
 
 export default function reactSvgLoadPlugin(options: PluginOpts = {}): any {
@@ -20,10 +21,14 @@ export default function reactSvgLoadPlugin(options: PluginOpts = {}): any {
 
       const contents = fs.readFileSync(id);
 
+      const componentName = options.componentName
+        ? options.componentName(id)
+        : null;
+
       return Promise.resolve(String(contents))
         .then(optimize(options.svgo))
-        .then(transform({ jsx: options.jsx }))
+        .then(transform({ jsx: options.jsx, componentName }))
         .then((result: any) => result.code);
-    }
+    },
   };
 }